finished breakpoints in the debugger

+++ src/serialize.h
@@ -7,7 +7,7 @@
 
 typedef struct {
     char name[32];
-    uint32_t startIdx;
+    uint64_t startIdx;
     uint32_t numArgs;
 } CometSerializedFunc;
